fix: rename the data file
